# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sovren API
  description: |
    **Sovren API**: NOSTR-native decentralized creator monetization platform

    ## Overview
    The Sovren API provides comprehensive endpoints for content management, user profiles,
    and payment processing via Bitcoin Lightning Network. Built on the NOSTR protocol
    for true decentralization and censorship resistance.

    ## Base URLs
    - **Development**: `http://localhost:3001/api/v1`
    - **Production**: `https://api.sovren.app/v1`

    ## Authentication
    All authenticated endpoints require a JWT bearer token obtained via the authentication flow.
    The token must include the user's NOSTR public key for protocol compliance.

    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## Rate Limiting
    API requests are rate-limited per endpoint category:
    - Content operations: 100 requests/hour
    - User operations: 200 requests/hour
    - Payment operations: 50 requests/hour

    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Total requests allowed
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Time when limit resets (Unix timestamp)

    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes:
    - `400`: Bad Request - Validation error
    - `401`: Unauthorized - Authentication required
    - `403`: Forbidden - Insufficient permissions
    - `404`: Not Found - Resource doesn't exist
...

...

...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /content/search:
    get:
      tags: [Content]
      summary: Search content
      description: |
        Performs full-text search across published content with advanced filtering and sorting options.
        Supports filtering by content type, tags, author, date range, and price range.
      operationId: searchContent
      parameters:
        - name: query
          in: query
          required: true
          description: Search query string
          schema:
            type: string
            minLength: 1
            maxLength: 500
          example: "bitcoin lightning network"
        - name: contentType
          in: query
          description: Filter by content type (comma-separated)
          schema:
            type: string
          example: "article,video"
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
          example: "bitcoin,nostr"
        - name: authorId
          in: query
          description: Filter by author user ID
          schema:
            type: string
            format: uuid
        - name: dateFrom
          in: query
          description: Filter by start date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          description: Filter by end date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: minPrice
          in: query
          description: Minimum price in sats
          schema:
            type: integer
            minimum: 0
        - name: maxPrice
          in: query
          description: Maximum price in sats
          schema:
            type: integer
            minimum: 0
        - name: sort
          in: query
          description: Sort order
          schema:
            type: string
            enum: [relevance, date, popularity, price]
            default: relevance
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Results per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchContentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /content/analytics/{id}:
    get:
      tags: [Content]
      summary: Get content analytics
      description: |
        Retrieves comprehensive analytics for a specific piece of content including views,
        engagement, revenue, traffic sources, and performance metrics.
      operationId: getContentAnalytics
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Content ID
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          description: Start date for analytics (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: End date for analytics (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: metrics
          in: query
          description: Specific metrics to retrieve (comma-separated)
          schema:
            type: string
          example: "views,engagement,revenue"
      responses:
        '200':
          description: Content analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentAnalyticsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from the authentication endpoint.

        Token must include:
        - User ID
        - NOSTR public key
        - User role (supporter, creator, admin)
        - Expiration timestamp

        Example:
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        ```
  headers:
    X-RateLimit-Limit:
      description: Total number of requests allowed in the current window
      schema:
        type: integer
      example: 100
...
  schemas:
...
    SearchContentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            results:
              type: array
              items:
                $ref: '#/components/schemas/ContentSearchResult'
            totalResults:
              type: integer
              example: 142
            currentPage:
              type: integer
              example: 1
            totalPages:
              type: integer
              example: 8
            searchTime:
              type: number
              format: float
              description: Search execution time in milliseconds
              example: 23.5
            facets:
              type: object
              description: Search result facets for filtering
              properties:
                contentTypes:
                  type: object
                  additionalProperties:
                    type: integer
                  example:
                    article: 89
                    video: 42
                    audio: 11
                tags:
                  type: object
                  additionalProperties:
                    type: integer
                  example:
                    bitcoin: 67
                    nostr: 45
                    lightning: 30
                priceRanges:
                  type: object
                  additionalProperties:
                    type: integer
                  example:
                    free: 78
                    low: 45
                    medium: 15
                    high: 4
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
...
    ContentAnalyticsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            contentId:
              type: string
              format: uuid
            timeRange:
              type: object
              properties:
                start:
                  type: string
                  format: date-time
                end:
                  type: string
                  format: date-time
            overview:
              type: object
              properties:
                totalViews:
                  type: integer
                  example: 1523
                uniqueViewers:
                  type: integer
                  example: 892
                totalEngagements:
                  type: integer
                  example: 234
                averageEngagementRate:
                  type: number
                  format: float
                  example: 0.154
                revenue:
                  type: object
                  properties:
                    totalSats:
                      type: integer
                      example: 125000
                    totalUSD:
                      type: number
                      format: float
                      example: 45.50
            engagement:
              type: object
              properties:
                likes:
                  type: integer
                shares:
                  type: integer
                comments:
                  type: integer
                saves:
                  type: integer
            traffic:
              type: object
              properties:
                sources:
                  type: object
                  additionalProperties:
                    type: integer
                  example:
                    direct: 450
                    nostr: 723
                    twitter: 215
                    search: 135
                devices:
                  type: object
                  additionalProperties:
                    type: integer
                  example:
                    mobile: 876
                    desktop: 587
                    tablet: 60
                geoLocations:
                  type: object
                  additionalProperties:
                    type: integer
                  example:
                    US: 623
                    UK: 298
                    CA: 187
            performance:
              type: object
              properties:
                averageReadTime:
                  type: integer
                  description: Average read time in seconds
                  example: 245
                completionRate:
                  type: number
                  format: float
                  example: 0.68
                bounceRate:
                  type: number
                  format: float
                  example: 0.32
            trends:
              type: array
              items:
                type: object
                properties:
                  timestamp:
                    type: string
                    format: date-time
                  views:
                    type: integer
                  engagement:
                    type: integer
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
...
  responses:
    ValidationError:
      description: Validation error - Request contains invalid data
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                success: false
                error:
                  code: "E001"
                  message: "Validation failed"
                  validationErrors:
                    - field: "title"
                      message: "Title is required"
                    - field: "amount"
                      message: "Amount must be greater than 0"
                metadata:
                  requestId: "req-123"
                  timestamp: "2025-10-27T10:30:00Z"
                  processingTime: 5.2
...
    RateLimitExceeded:
      description: Too Many Requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          description: Seconds until rate limit resets
          schema:
            type: integer
          example: 3600
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limit:
              value:
                success: false
                error:
                  code: "E429"
                  message: "Rate limit exceeded"
                  details:
                    limit: 100
                    window: "1 hour"
                    retryAfter: 3600
...